requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.4
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2